requirements.txt
Flask==2.0.1
tellurium==2.2.0
pymongo==3.12.0
dnspython==2.1.0
numpy==1.21.4